+++ packages/user-onboarding/projects/user-onboarding-lib/src/lib/user-onboarding-lib.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+
+import { StorageServiceModule } from 'ngx-webstorage-service';
+import {VgCoreModule} from '@videogular/ngx-videogular/core';
+import {VgControlsModule} from '@videogular/ngx-videogular/controls';
+import {VgOverlayPlayModule} from '@videogular/ngx-videogular/overlay-play';
+import {VgBufferingModule} from '@videogular/ngx-videogular/buffering';
+
+@NgModule({
+  declarations: [
+
+  ],
+  imports: [
+    StorageServiceModule,
+    VgCoreModule,
+    VgControlsModule,
+    VgOverlayPlayModule,
+    VgBufferingModule
+  ],
+  exports: [
+  ]
+})
+export class UserOnboardingLibModule { }