+++ packages/user-onboarding/projects/user-onboarding-lib/src/models/StateParameters.ts
@@ -1,0 +1,11 @@
+import { TourState } from './Tour';
+
+export interface SaveStateParameters {
+  tourId: string;
+  state: TourState;
+}
+
+export interface LoadStateParameters {
+  tourId: string;
+  sessionId: string;
+}