+++ packages/user-onboarding/projects/user-onboarding-lib/src/models/TourParameters.ts
@@ -1,0 +1,11 @@
+import { TourStep } from './TourStep';
+
+export interface SaveTourParameters {
+  tourId: string;
+  tourSteps: TourStep[];
+  styleSheet: string;
+}
+
+export interface LoadTourParameters {
+  tourId: string;
+}