+++ packages/user-onboarding/projects/user-onboarding-lib/src/models/index.ts
@@ -1,0 +1,7 @@
+export * from './StateParameters';
+export * from './Tour';
+export * from './TourButton';
+export * from './TourParameters';
+export * from './TourStep';
+export * from './parameterModel';
+export * from './videoModel';