+++ packages/user-onboarding/projects/user-onboarding-lib/src/models/parameterModel.ts
@@ -1,0 +1,9 @@
+import { HttpHeaders, HttpParams } from '@angular/common/http';
+
+export interface ParameterModel {
+  UUID: string;
+  headers?: HttpHeaders;
+  // observe?: HttpObserve;
+  query?: HttpParams;
+  reportProgress?: boolean;
+}