+++ services/authentication-service/src/application.ts
@@ -1,0 +1,27 @@
+import {BootMixin} from '@loopback/boot';
+import {ApplicationConfig} from '@loopback/core';
+import {RepositoryMixin} from '@loopback/repository';
+import {RestApplication} from '@loopback/rest';
+import * as path from 'path';
+import {AuthenticationServiceComponent} from './component';
+
+export {ApplicationConfig};
+
+export class AuthenticationServiceApplication extends BootMixin(
+  RepositoryMixin(RestApplication),
+) {
+  constructor(options: ApplicationConfig = {}) {
+    super(options);
+    this.static('/', path.join(__dirname, '../public'));
+    this.component(AuthenticationServiceComponent);
+
+    this.projectRoot = __dirname;
+    this.bootOptions = {
+      controllers: {
+        dirs: ['controllers'],
+        extensions: ['.controller.js'],
+        nested: true,
+      },
+    };
+  }
+}