+++ services/authentication-service/src/modules/auth/login.controller.ts
@@ -311,6 +311,7 @@ export class LoginController {
   async exchangeToken(
     @requestBody() req: AuthRefreshTokenRequest,
     @param.header.string('device_id') deviceId?: string,
+    @param.header.string('Authorization') token?: string,
   ): Promise<TokenResponse> {
     const refreshPayload: RefreshToken = await this.refreshTokenRepo.get(
       req.refreshToken,
@@ -325,6 +326,10 @@ export class LoginController {
     });
     if (!authClient) {
       throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
+    }
+    const accessToken = token?.split(' ')[1];
+    if (!accessToken || refreshPayload.accessToken !== accessToken) {
+      throw new HttpErrors.Unauthorized(AuthErrorKeys.TokenInvalid);
     }
     await this.revokedTokensRepo.set(refreshPayload.accessToken, {
       token: refreshPayload.accessToken,
@@ -653,10 +658,6 @@ export class LoginController {
       },
     },
   })
-
-  /**
-   * @deprecated This method should not be used, possible security issue if secret is passed via query params, please use the post endpoint
-   */
   async loginViaKeycloak(
     @param.query.string('client_id')
     clientId?: string,